Moved MIME related code to its own file.
